+++ tests/testthat/test_set.R
@@ -109,7 +109,7 @@ test_that("isSubset",{
 })
 
 test_that("as.Set",{
-  expect_equal(as.Set(c(1,2)), Set$new(1,2))
+  expect_equal(as.Set(c(1,2)), Set$new(1,2,class = "numeric"))
   expect_equal(as.Set(list(1,2)), list(Set$new(1), Set$new(2)))
   expect_equal(as.Set(matrix(c(1,2,3,4),nrow = 2)), list(Set$new(1,2),Set$new(3,4)))
   expect_error(expect_equal(as.Set(Interval$new()), Interval$new()), "Interval cannot be")