+++ tests/testthat/test_set_universalset.R
@@ -1,7 +1,5 @@
 library(testthat)
 
-context("UniversalSet")
-
 test_that("construction",{
   expect_silent(UniversalSet$new())
   expect_error(UniversalSet$new(12))